+++ CHANGELOG.md
@@ -4,7 +4,22 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
 
 ## Next
 
+### Bug Fixes
+
+* scheduler: Fix regression causing controller names of less than 10 characters
+  to fail to register jobs
+  ([PR](https://github.com/hashicorp/boundary/pull/2226)).
+* sessions: Fix an additional case from the changes in the 0.8.x series that
+  could result in sessions never moving from `canceling` state to terminated.
+  ([PR](https://github.com/hashicorp/boundary/pull/2229))
+
 ## 0.9.0 (2022/06/20)
+
+### Known Issues
+
+* If a controller's defined name in a configuration file is less than 10
+  characters, errors may be seen on startup related to registration of jobs.
+  This is a regression in this version and will be fixed in the next release.
 
 ### New and Improved